+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 NEW FEATURES:
 
-  - Triggers can now be run as a provisioner [(#21)](https://github.com/emyl/vagrant-triggers/issues/21)
+  - Actions can be blacklisted using ```config.trigger.blacklist```.
+  - Triggers can be run as a provisioner [(#21)](https://github.com/emyl/vagrant-triggers/issues/21)
 
 ## 0.4.3 (November 26, 2014)