+++ CHANGELOG.md
@@ -1,7 +1,12 @@
 ## 0.5.0 (unreleased)
 
+**BEHAVIOURAL CHANGES:**
+
+  - The ```:stdout``` option now defaults to true.
+
 NEW FEATURES:
 
+  - New option ```:stderr``` for displaying standard error from scripts.
   - The special action ```:ALL``` can be used when a trigger should always run [(#23)](https://github.com/emyl/vagrant-triggers/issues/23)
   - Actions can be blacklisted using ```config.trigger.blacklist```.
   - Triggers can be run as a provisioner [(#21)](https://github.com/emyl/vagrant-triggers/issues/21)